fixed LED config, and set colour to white
for some reason the default config for this keyboard gives the layer indicator LED a flag of 'NONE', which is... A Decision. it means the LED wont turn on at all ever. makes things difficult.
I also set the LED colour to white so i could verify the LEDs work